Use Bresenham's to render triangle edges.

+++ lib/os/os.h
@@ -19,7 +19,7 @@
 #endif
 
 #ifndef bool
-typedef int bool;
+typedef uint8_t bool;
 #define true 1
 #define false 0
 #endif
